requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
pyqrcode==1.2.1
Pillow==10.0.1
python-dotenv==1.0.0
